+++ R/4-propd.R
@@ -106,8 +106,9 @@ propd <- function(counts, group, alpha, p = 100, weighted = FALSE){
 
   # Replace zeros unless alpha is provided
   if(any(as.matrix(counts) == 0) & is.na(alpha)){
-    message("Alert: Replacing 0s in \"count matrix\" with 1.")
-    ct[ct == 0] <- 1
+    message("Alert: Replacing 0s with next smallest value.")
+    zeros <- ct == 0
+    ct[zeros] <- min(ct[!zeros])
   }
 
   # Initialize @active, @weighted
@@ -214,5 +215,11 @@ setDisjointed <- function(propd){
 #'  A wrapper for \code{setActive(propd, what = "theta_e")}.
 #' @export
 setEmergent <- function(propd){
+
+  if(table(propd@group)[1] != table(propd@group)[2]){
+
+    warning("Emergent proportionality not yet validated for unequal group sizes.")
+  }
+
   setActive(propd, what = "theta_e")
 }