+++ R/5-propdCall.R
@@ -100,6 +100,13 @@ calculateTheta <- function(counts, group, alpha, lrv = NA, only = "all",
     if(only == "theta_f") return(theta_f)
   }
 
+  if(only == "all" | only == "theta_g"){
+
+    theta_g <- pmin(p1 * lrv1, p2 * lrv2) / (p * lrv)
+    if(replaceNaNs) theta_g[lrv0] <- 1
+    if(only == "theta_g") return(theta_g)
+  }
+
   labels <- labRcpp(ncol(counts))
   return(
     data.frame(
@@ -108,6 +115,7 @@ calculateTheta <- function(counts, group, alpha, lrv = NA, only = "all",
       "theta" = theta,
       "theta_e" = theta_e,
       "theta_f" = theta_f,
+      "theta_g" = theta_g,
       "lrv" = lrv,
       "lrv1" = lrv1,
       "lrv2" = lrv2,
@@ -258,7 +266,7 @@ updateF <- function(propd, moderated = FALSE, ivar = "clr"){
 
     propd@Fivar <- NA # used by updateCutoffs
     Fstat <- (n1 + n2 - 2) * (1 - propd@results$theta) / propd@results$theta
-    theta_mod <- 0
+    theta_mod <- as.numeric(NA)
   }
 
   propd@results$theta_mod <- theta_mod
@@ -268,6 +276,56 @@ updateF <- function(propd, moderated = FALSE, ivar = "clr"){
   K <- length(unique(propd@group))
   N <- n1 + n2 + propd@dfz
   propd@results$Pval <- pf(Fstat, K - 1, N - K, lower.tail = FALSE)
+  propd@results$FDR <- p.adjust(propd@results$Pval, method = "BH")
 
   return(propd)
+}
+
+#' Calculate a theta Cutoff
+#'
+#' This function uses the F distribution to calculate a cutoff of
+#'  theta for a p-value given by the \code{pval} argument.
+#'
+#' @inheritParams all
+#' @param pval A p-value at which to calculate a theta cutoff.
+#'
+#' @return A cutoff of theta from [0, 1].
+#'
+#' @export
+qtheta <- function(propd, moderated = FALSE, pval = 0.05){
+
+  if(pval < 0 | pval > 1) stop("Provide a p-value cutoff from [0, 1].")
+
+  K <- length(unique(propd@group))
+  N <- length(propd@group)
+
+  if(moderated){
+
+    propd <- suppressMessages(updateF(propd, moderated = TRUE))
+    z.df <- propd@dfz
+
+    Q <- qf(pval, K - 1, N + z.df - K, lower.tail = FALSE)
+    # # Fstat <- (n1 + n2 + z.df - 2) * Fprime
+    # # theta_mod <- 1 / (1 + Fprime)
+    # # Q = Fstat
+    # # Q = (n1 + n2 + z.df - 2) * Fprime
+    # # Fprime = 1/theta_mod - 1
+    R <- N - 2 + z.df
+    # # Q = R * (1/theta_mod - 1)
+    # # Q = R/theta_mod - R
+    theta_a05 <- R/(Q+R)
+
+  }else{
+
+    Q <- qf(pval, K - 1, N - K, lower.tail = FALSE)
+    # # Fstat <- (N - 2) * (1 - propd@theta$theta) / propd@theta$theta
+    # # Q = Fstat
+    # # Q = (N-2) * (1-theta) / theta
+    # # Q / (N-2) = (1/theta) - 1
+    # # 1/theta = Q / (N-2) + 1 = Q(N-2)/(N-2)
+    # # theta = (N-2)/(Q+(N-2))
+    theta_a05 <- (N-2)/(Q+(N-2))
+  }
+
+  return(theta_a05)
 }